+++ apps/pioneer-template/src/utils/onStart.tsx
@@ -1,0 +1,35 @@
+// src/utils/onStartApp.tsx
+import { usePioneer } from "@coinmasters/pioneer-react";
+import { WalletOption, availableChainsByWallet } from '@coinmasters/types';
+
+export const useOnStartApp = () => {
+    const { onStart } = usePioneer();
+
+    const onStartApp = async () => {
+        try {
+            let walletsVerbose = [];
+            const { keepkeyWallet } = await import("@coinmasters/wallet-keepkey");
+
+            const pioneerSetup: any = {
+                appName: "KeepKey Vault",
+                appIcon: "https://i.imgur.com/0UxjLIF.png",
+            };
+
+            const walletKeepKey = {
+                type: WalletOption.KEEPKEY,
+                icon: "https://pioneers.dev/coins/keepkey.png",
+                chains: availableChainsByWallet[WalletOption.KEEPKEY],
+                wallet: keepkeyWallet,
+                status: "offline",
+                isConnected: false,
+            };
+
+            walletsVerbose.push(walletKeepKey);
+            onStart(walletsVerbose, pioneerSetup);
+        } catch (e) {
+            console.error("Failed to start app!", e);
+        }
+    };
+
+    return onStartApp;
+};