+++ packages/swapkit/api/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @swapkit/api
 
+## 3.7.35
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.7.35
+
 ## 3.7.34
 
 ### Patch Changes