+++ packages/toolboxes/cosmos/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @coinmasters/toolbox-cosmos
 
+## 11.0.35
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.7.35
+  - @coinmasters/api@3.7.35
+
 ## 11.0.34
 
 ### Patch Changes