+++ packages/toolboxes/utxo/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @swapkit/toolbox-utxo
 
+## 11.0.35
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.7.35
+
 ## 11.0.34
 
 ### Patch Changes